src/components/films-details/films-details.spec.ts
import { FilmsDetails } from './films-details';
import { newSpecPage } from '@stencil/core/testing';

describe('films-details', () => {
  describe('normalization', () => {
    it('returns a blank string if the name is undefined', async () => {
      const { rootInstance } = await newSpecPage({
        components: [FilmsDetails],
        html: '<films-details></films-details>',
      });
      expect(rootInstance.normalize(undefined)).toEqual('');
    });

    it('returns a blank string if the name is null', async () => {
      const { rootInstance } = await newSpecPage({
        components: [FilmsDetails],
        html: '<films-details></films-details>',
      });
      expect(rootInstance.normalize(null)).toEqual('');
    });

    it('capitalizes the first letter', async () => {
      const { rootInstance } = await newSpecPage({
        components: [FilmsDetails],
        html: '<films-details></films-details>',
      });
      expect(rootInstance.normalize('quincy')).toEqual('Quincy');
    });

    it('lower-cases the following letters', async () => {
      const { rootInstance } = await newSpecPage({
        components: [FilmsDetails],
        html: '<films-details></films-details>',
      });
      expect(rootInstance.normalize('JOSEPH')).toEqual('Joseph');
    });

    it('handles single letter names', async () => {
      const { rootInstance } = await newSpecPage({
        components: [FilmsDetails],
        html: '<films-details></films-details>',
      });
      expect(rootInstance.normalize('q')).toEqual('Q');
    });
  });
});
